backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
geoalchemy2>=0.15.0
alembic>=1.14.0
earthengine-api>=1.4.0
httpx>=0.28.0
shapely>=2.0.6
rasterio>=1.4.0
numpy>=1.26.0
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
